Merge the delta_content attribute into delta_status
Previously, delta_content would show if the resulting binary was changed
or not. However, this was only relevant when both the base and
comparison build had status OK, and the attribute would be confusingly
empty in other cases.

This effectively splits the "Unchanged" status into "Identical",
"Modified", "Still Broken" and "Is Base". Additionally, a status of "No
Base" is added when no  corresponding base build was found, so each
build should now have a delta_status.