Make building work for the sam core
Previously, determining compile sizes used avr-size directly, but this
obviously doesn't work on sam (ARM). Now, the Arduino IDE compiler
output is parsed (though that will break when the IDE is not configured
for English, but well).

On the longer term, the IDE should probably be modified to allow
outputting sizes in a language-independent and easily parsable manner,
or it should easily expose the recipes from platform.txt so we can call
the right commands (without hardcoding) ourselves.

Additionally, this uses the .bin file instead of the .hex file when the
latter doesn't exist - the sam core produces .bin files while the avr
core produces .hex files.